requirements.txt
django==5.0.6
djangorestframework==3.15.2
psycopg2-binary==2.9.9
python-decouple==3.8
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
drf-spectacular==0.27.2
django-filter==24.2
gunicorn==22.0.0
whitenoise==6.7.0
sentry-sdk[django]==2.3.1
bleach==6.1.0
# dev / test
pytest==8.2.0
pytest-django==4.8.0
pytest-cov==5.0.0
factory-boy==3.3.0
faker==24.11.0
ruff==0.3.7
black==24.4.2
